First try to fix Phone
On 2.6.32 we did fixed phone with some patches only related to
the usb drivers. So i comes to my mind since every 2ndboot-patch
is already applied to our kernel source that this might be the issue.

We did prevent TLL soft reset and i did the same here, my intention is
that the actual "fix" is somewhere else in the code.

This commit will result in a bootloop, because of some merging issues
(i tried to revert as much as i can from our kernel to bring it to a
stage where tll soft reset was inside).

Maybe this helps someone, maybe its the solution or maybe its totaly
useless anyway...