mmc: core: sdio: added core function of suspend/resume
This is my own implementation of suspend/resume for
our new wifi driver. The idea behind is to enable
wifi during screen on and disable it while screen is off.
It's much simplier than the orginal one.

In the current state it won't work correctly, as soon as
it gets into sleep it reboots (see FIXME). Also because of
mmc_stop_host the firmware won't be generated on each boot
and has to be added manually.

Change-Id: I6934ceee471192f9ae513f5c2da3d6b793e1e4d8